requirement.txt
django
djangorestframework
django-cors-headers
opencv-python
Pillow 
fer 
google-generativeai 
python-dotenv 
boto3
moviepy
pandas